// src/shared/domain/enums/Integration-config-key.enum.ts
export enum IntegrationConfigKey {
    COLUMNS_MAPPING = 'columns_mapping',
    PROJECT_MANAGEMENT_SETUP_CONFIG = 'project_management_setup_config',
    REPOSITORIES = 'repositories',
    INSTALLATION_GITHUB = 'installation_github',
    CHANNEL_INFO = 'channel_info',
    MSTEAMS_INSTALLATION_APP = 'msteams_installation_app',
    WAITING_COLUMNS = 'waiting_columns',
    DOING_COLUMN = 'doing_column',
    DAILY_CHECKIN_SCHEDULE = 'daily_checkin_schedule',
    MODULE_WORKITEMS_TYPES = 'module_workitems_types',
    BUG_TYPE_IDENTIFIERS = 'bug_type_identifier',
    AUTOMATION_ISSUE_ALERT_TIME = 'automation_issue_alert_time',
    TEAM_PROJECT_MANAGEMENT_METHODOLOGY = 'team_project_management_methodology',
    CODE_MANAGEMENT_PAT = 'code_management_pat',
    USE_JQL_TO_VIEW_BOARD = 'use_jql_to_view_board',
}

export enum IntegrationConfigKeyProjectManagement {
    USE_JQL_TO_VIEW_BOARD = 'use_jql_to_view_board',
    MODULE_WORKITEMS_TYPES = 'module_workitems_types',
    BUG_TYPE_IDENTIFIERS = 'bug_type_identifier',
    AUTOMATION_ISSUE_ALERT_TIME = 'automation_issue_alert_time',
    DAILY_CHECKIN_SCHEDULE = 'daily_checkin_schedule',
    TEAM_PROJECT_MANAGEMENT_METHODOLOGY = 'team_project_management_methodology',
    WAITING_COLUMNS = 'waiting_columns',
    DOING_COLUMN = 'doing_column',
}